Refactor BlogArchive component to implement pagination and enhance post visibility logic
