+++ .vitepress/theme/components/PostList.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { computed, onMounted, ref, watch } from "vue"
+import { onMounted, ref } from "vue"
 
 import { formatDate, type Content } from "../lib/content"
 import styles from "./PostList.module.css"
@@ -7,70 +7,6 @@ import styles from "./PostList.module.css"
 const props = defineProps<{
   posts: Content[]
 }>()
-
-const pageSize = 20
-const currentPage = ref(1)
-const pageQueryParam = "page"
-
-const totalPages = computed(() => {
-  return Math.max(1, Math.ceil(props.posts.length / pageSize))
-})
-
-const pagedPosts = computed(() => {
-  const start = (currentPage.value - 1) * pageSize
-  return props.posts.slice(start, start + pageSize)
-})
-
-watch(
-  () => props.posts,
-  () => {
-    const pageFromUrl = getPageFromUrl()
-    currentPage.value = pageFromUrl ?? 1
-  },
-)
-
-function goToPage(page: number) {
-  currentPage.value = Math.min(Math.max(1, page), totalPages.value)
-}
-
-function getPageFromUrl() {
-  if (typeof window === "undefined") {
-    return null
-  }
-
-  const rawValue = new URLSearchParams(window.location.search).get(pageQueryParam)
-  if (!rawValue) {
-    return null
-  }
-
-  const parsed = Number(rawValue)
-  if (!Number.isInteger(parsed) || parsed < 1) {
-    return null
-  }
-
-  return parsed
-}
-
-function syncPageToUrl(page: number) {
-  if (typeof window === "undefined") {
-    return
-  }
-
-  const url = new URL(window.location.href)
-
-  if (page <= 1) {
-    url.searchParams.delete(pageQueryParam)
-  } else {
-    url.searchParams.set(pageQueryParam, String(page))
-  }
-
-  const nextUrl = `${url.pathname}${url.search}${url.hash}`
-  const currentUrl = `${window.location.pathname}${window.location.search}${window.location.hash}`
-
-  if (nextUrl !== currentUrl) {
-    window.history.replaceState(window.history.state, "", nextUrl)
-  }
-}
 
 const visitedPostUrls = ref<Set<string>>(new Set())
 const visitedStorageKey = "hs-visited-posts"
@@ -117,30 +53,12 @@ onMounted(() => {
   } catch {
     window.localStorage.removeItem(visitedStorageKey)
   }
-
-  const pageFromUrl = getPageFromUrl()
-  if (pageFromUrl) {
-    currentPage.value = Math.min(pageFromUrl, totalPages.value)
-  }
 })
-
-watch(
-  [currentPage, totalPages],
-  () => {
-    if (currentPage.value > totalPages.value) {
-      currentPage.value = totalPages.value
-      return
-    }
-
-    syncPageToUrl(currentPage.value)
-  },
-)
 </script>
 
 <template>
-  <div>
-    <div :class="styles.postList">
-      <article v-for="post in pagedPosts" :key="post.url" :class="styles.postItem">
+  <div :class="styles.postList">
+    <article v-for="post in props.posts" :key="post.url" :class="styles.postItem">
       <a
         :class="[styles.postLink, !post.date && !post.image && styles.postLinkNoMeta]"
         :href="post.url"
@@ -187,30 +105,5 @@ watch(
         </span>
       </a>
     </article>
-    </div>
-
-    <nav v-if="totalPages > 1" :class="styles.pagination" aria-label="Post list pagination">
-      <button
-        type="button"
-        :class="styles.paginationButton"
-        :disabled="currentPage === 1"
-        @click="goToPage(currentPage - 1)"
-      >
-        Previous
-      </button>
-
-      <span :class="styles.paginationLabel">
-        Page {{ currentPage }} of {{ totalPages }}
-      </span>
-
-      <button
-        type="button"
-        :class="styles.paginationButton"
-        :disabled="currentPage === totalPages"
-        @click="goToPage(currentPage + 1)"
-      >
-        Next
-      </button>
-    </nav>
   </div>
 </template>
